Clean up requirements
Paramiko used to be required by ironic, now it's only used by ansible
and should be handled by its requirements.

Simplify the setuptools requirement to use the version from
Bionic as a minimum and pull in the latest exclusions from g-r.

Move PyYAML to requirements since it's used in production code.

Remove test requirements that are not referenced anywhere.

Change-Id: Idc9f1cd66b123c9c9325d7fd7a557475bfdbdd68