Clean up bindep.txt
Since we have other means of installing packages, bindep should only
be used for the packages required to bootstrap bifrost, i.e. python
and development files. Things like MariaDB should wait.

ebtables and dnsmasq is needed for libvirt on debian systems, so
install it there (and in suse since it's broken now and I cannot test).

Change-Id: I955f559e247ff49f5b84453a8c48462c63413acd