Fix missing node facts with BIFROST_INVENTORY_SOURCE=ironic
In Ussuri and earlier releases, the Bifrost custom inventory returns all
node fields as facts. Currently, most of these are returned as null.
This happened following the switch to openstacksdk. The plugin checks
for the presence of a properties field, and gets detailed node
information if it is missing. This worked with shade, but with
openstacksdk, the returned node object has most fields present but set
to null.

One of the most obvious consequences is that provisioning fails, since
the 'Deploy to hardware - Using custom instance_info.' task tries to use
the instance_info value of null.

This change fixes the issue by removing the check and always getting the
detailed node information.

Change-Id: Ia87c8332994f3b0f037ada953a299987bba246e5
Story: 2008394
Task: 41321