Revert "Explicitly trap on ERR" and fix exit code hanlding
Trapping on ERR does not do what is implied here and is redundant when
trapping on EXIT is used together with `set -e`. The real problem is
that in the CI scripts we pipe test-bifrost.sh into tee with a log file
without setting -o pipefail in bash. This change fixes it.

This reverts commit b31bc66726ddcff11678a77cce8b14f629a1b19d.

Change-Id: Ifca4049f5967c0b2153ecffb4c153a96b4a62cc7