Move /httpboot to /var/lib/ironic
There is no guarantee that /var is not on a separate filesystem. Also
we should not pollute the root directory with our stuff.

The TFTP directory /tftpboot is left where it is, since it needs to be
accessed by dnsmasq. We may consider moving it later on.

Removes the wildly outdated statement about purging the image cache
from the documentation.

Change-Id: Ib1e46e7a9c5eec193082858614026ca4c9f537ac