Allow for almalinux and rocky installations
With the changes to bifrost-ironic-install required_defaults for
the RedHat_family, we can now try to not get in the way of
deploying on distros like almalinux and rocky.

Fedora 32 has been EOL for close to a year, so we'll just assume
firewalld for that case. Distributions RedHat, CentOS and Fedora
are in the RedHat_family, so assume firewalld all distributions
in the family.

Change-Id: I1c8b768a3daf0af0d8e30d935558b6fdd912adfd
(cherry picked from commit 7640a231a84f87939d3242b062f4ca1e89bcfa4a)