Fix TFTP read access for enforcing SELinux
This change adds ironic_tftp_master_path to the list of paths to get
the tftpdir_t context, so that images copied from here to
tftp_boot_folder will have a context which allows the file to be read.

This change also applies the context changes when SELinux is in
Permissive mode, as well as Enforcing.

Change-Id: I43dd4a2f7cf47934bc9f5cdee85acf53a40dd468