Refactor use of include_vars
Move all these os-default files to vars directory, where the
include_vars expects them to be.

Use a first_found lookup instead of a with_first_found loop so that
the 'paths' parameter can be used.

This ensures that only vars from the role are included, and not vars
from a parent calling role. This can happen when a parent role has
a higher priority vars file available for inclusion than the role
it calls.

The lookup order was taken from the openstack-ansible project, and
can be extended by copying the family vars file to a more specific
distro or versioned file and making necessary the changes, should
the need to complicate things ever arise again.

Change-Id: Ic3c0a072fc8d4cb10f6431c118f14afba3a3da1f