Refactor the use of include_vars
The debian_family and ubuntu files have the same content, and they're
both in the debian family, so drop the ubuntu vars.

Move remaining os-default files to vars directory, where the
include_vars expects them to be.

Use a first_found lookup instead of a with_first_found loop so that
the 'paths' parameter can be used.

This ensures that only vars from the role are included, and not vars
from a parent calling role. This can happen when a parent role has
a higher priority vars file available for inclusion than the role
it calls.

Change-Id: I93b0d7f296bae859914477b4009e44a44e3bde1a