ironic: Perform online data migrations with localhost DB
Previously online data migrations were not performed in the default
configuration where the database address is 'localhost'. I'm not sure
why this condition was used, but it was present in the original
change I685efd14bf3567a126311b676a50b0abb0f043db submitted adding the
online data migrations back in the Stein release.

This change removes the condition to perform online data migrations
regardless of the database address.

Change-Id: I60d0f61a3c6fd3573b2031f7dba5d8a29e178394
Closes-Bug: #2036772