Configure static DHCP before automated cleaning
After the switch to newer ansible collection, the enrollment workflow
started running automated cleaning (before that, it would use the old
API version resulting in immediate "available"). However, the static
DHCP configuration only happened in the inspect and deploy workflows,
which are run after enrollment.

This change extracts a new small role for the DHCP config and includes
it in all 3 workflows that use DHCP.

While here, make sure that dnsmasq_dhcp_hostsdir is respected.

Change-Id: Idf6f24dde11d600698d45a218812cba8134fb73f