Fix checksum parsing for Cirros and Rocky Linux
For Rocky Linux, the checksum file format differs from Ubuntu.  The
current awk expression failed to extract the correct checksum and would
instead output SHA256. This caused checksum validation to fail. A
similar issue exists for CirrOS which uses the MD5 algorithm.

Changes the logic to first extract the line with the corresponding
image name and then searches for a string that looks like a checksum
within that line.

Closes-Bug: #2081031
Change-Id: Id426db0ad418898a2ebe9f2b5001945520dc6b1d