Use python version to set DEFAULT_PIP_ANSIBLE
Instead of assuming the python version based on the distro, we
check the python version. (Check borrowed from bifrost-cli)

This patch defaults to the newer ansible version, and if python
is old, we override it with the old ansible package.

This lets us install on the RHELatives like Alma and Rocky, which
the current check is preventing.

Change-Id: I47413eab8c5132ab3201b4c7197ec15c17d18e12