Centos: set masked: false for firewalld
It appears that Centos has changed the default firewall masking state
and as such the playbook now fails. Adding masked: false to unmask the
unit so hopefuly bifrost works as expected.

Also, temporarily disables the bookworm job so we can un-wedge CI
which is broken in multiple ways including centos breakages.

Change-Id: Ib30795a88d4908fd1b78250b99a840bccb185ba3
Signed-off-by: Julia Kreger <juliaashleykreger@gmail.com>